♻️[refactor] Separação dos arquivos em pacotes
