--- src/com/app/Entity.java
@@ -1,5 +1,0 @@
-package com.app;
-
-interface Entity {
-	public int getId();
-}
