+++ src/com/app/entities/Entity.java
@@ -1,0 +1,5 @@
+package com.app.entities;
+
+public interface Entity {
+	public int getId();
+}
