+++ src/com/app/util/Secret.java
@@ -1,4 +1,4 @@
-package com.app;
+package com.app.util;
 
 import java.io.File;
 import java.io.FileInputStream;
